Export MS Word content to JSON
This a small macro i wrote to export word content to JSON format. 
I am not a VBA programmer, I wanted it for one of my project and I couldn't get on the net. So I decided to write it
myself, I searched for pieces of code on net and customized it to my needs. Feel free to update/improve this code